npm build fixes, updates to answer generation, added markdown parsing packages
