+++ .eslintrc.json
@@ -1,0 +1,15 @@
+{
+  "extends": "next/core-web-vitals",
+  "rules": {
+    "@typescript-eslint/no-explicit-any": "off",
+    "@typescript-eslint/no-unused-vars": ["warn", { 
+      "argsIgnorePattern": "^_",
+      "varsIgnorePattern": "^_" 
+    }],
+    "react/no-unescaped-entities": "off",
+    "@next/next/no-img-element": "warn",
+    "prefer-const": "warn",
+    "import/no-anonymous-default-export": "off",
+    "react-hooks/exhaustive-deps": "warn"
+  }
+}
