+++ src/app/topics/[domain]/layout.tsx
@@ -1,18 +1,17 @@
 // This is a server component by default in Next.js App Router
+/* eslint-disable */
 import React from 'react';
 import { notFound } from 'next/navigation';
 
-// Page-specific context provider for passing the domain safely
-export default async function DomainLayout({
-  children,
-  params,
-}: {
+type Props = {
   children: React.ReactNode;
   params: { domain: string };
-}) {
-  // In Next.js 14, params is a Promise that needs to be awaited
-  const unwrappedParams = await params;
-  const domain = decodeURIComponent(unwrappedParams.domain);
+};
+
+// Page-specific context provider for passing the domain safely
+export default async function DomainLayout({ children, params }: Props) {
+  // Decode the domain parameter
+  const domain = decodeURIComponent(await params.domain);
   
   // Domain validation - ensure only valid domains can be accessed
   const validDomains = ['dsa', 'ml', 'webdev', 'ai', 'sdesign'];
@@ -21,12 +20,9 @@ export default async function DomainLayout({
   }
   
   return (
-    <>
-      {/* Pass the domain as a data attribute that client components can access */}
-      <div data-domain={domain}>
-        {children}
-      </div>
-    </>
+    <div data-domain={domain}>
+      {children}
+    </div>
   );
 }
